+++ lib/serega/config.rb
@@ -13,12 +13,15 @@ class Serega
     DEFAULTS = {
       plugins: [],
       initiate_keys: %i[only with except check_initiate_params].freeze,
-      attribute_keys: %i[method value serializer many hide const delegate default].freeze,
+      attribute_keys: %i[method value serializer many hide const delegate default preload preload_path].freeze,
       serialize_keys: %i[context many].freeze,
       check_attribute_name: true,
       check_initiate_params: true,
       delegate_default_allow_nil: false,
       max_cached_plans_per_serializer_count: 0,
+      auto_preload_attributes_with_delegate: false,
+      auto_preload_attributes_with_serializer: false,
+      auto_hide_attributes_with_preload: false,
       to_json: (SeregaJSON.adapter == :oj) ? SeregaJSON::OjDump : SeregaJSON::JSONDump,
       from_json: (SeregaJSON.adapter == :oj) ? SeregaJSON::OjLoad : SeregaJSON::JSONLoad
     }.freeze
@@ -157,6 +160,54 @@ class Serega
       # @return [#call] Provided callable object
       def from_json=(value)
         opts[:from_json] = value
+      end
+
+      # Returns :auto_preload_attributes_with_delegate config option
+      # @return [Boolean] Current :auto_preload_attributes_with_delegate config option
+      def auto_preload_attributes_with_delegate
+        opts.fetch(:auto_preload_attributes_with_delegate)
+      end
+
+      # Sets :auto_preload_attributes_with_delegate config option
+      #
+      # @param value [Boolean] Set :auto_preload_attributes_with_delegate config option
+      #
+      # @return [Boolean] :auto_preload_attributes_with_delegate config option
+      def auto_preload_attributes_with_delegate=(value)
+        raise SeregaError, "Must have boolean value, #{value.inspect} provided" if (value != true) && (value != false)
+        opts[:auto_preload_attributes_with_delegate] = value
+      end
+
+      # Returns :auto_preload_attributes_with_serializer config option
+      # @return [Boolean] Current :auto_preload_attributes_with_serializer config option
+      def auto_preload_attributes_with_serializer
+        opts.fetch(:auto_preload_attributes_with_serializer)
+      end
+
+      # Sets :auto_preload_attributes_with_serializer config option
+      #
+      # @param value [Boolean] Set :auto_preload_attributes_with_serializer config option
+      #
+      # @return [Boolean] :auto_preload_attributes_with_serializer config option
+      def auto_preload_attributes_with_serializer=(value)
+        raise SeregaError, "Must have boolean value, #{value.inspect} provided" if (value != true) && (value != false)
+        opts[:auto_preload_attributes_with_serializer] = value
+      end
+
+      # Returns :auto_hide_attributes_with_preload config option
+      # @return [Boolean] Current :auto_hide_attributes_with_preload config option
+      def auto_hide_attributes_with_preload
+        opts.fetch(:auto_hide_attributes_with_preload)
+      end
+
+      # Sets :auto_hide_attributes_with_preload config option
+      #
+      # @param value [Boolean] Set :auto_hide_attributes_with_preload config option
+      #
+      # @return [Boolean] :auto_hide_attributes_with_preload config option
+      def auto_hide_attributes_with_preload=(value)
+        raise SeregaError, "Must have boolean value, #{value.inspect} provided" if (value != true) && (value != false)
+        opts[:auto_hide_attributes_with_preload] = value
       end
     end