+++ spec/serega/attribute_normalizer_spec.rb
@@ -41,24 +41,6 @@ RSpec.describe Serega::SeregaAttributeNormalizer do
     end
   end
 
-  describe "#hide" do
-    it "returns provided :hide option" do
-      expect(normalizer.new(opts: {hide: true}).hide).to be true
-      expect(normalizer.new(opts: {hide: false}).hide).to be false
-      expect(normalizer.new(opts: {}).hide).to be_nil
-    end
-
-    it "returns saved :hide option" do
-      norm = normalizer.new(opts: {hide: true})
-
-      expect(norm.hide).to be true
-      expect(norm.instance_variable_get(:@hide)).to be true
-
-      norm.instance_variable_set(:@hide, nil)
-      expect(norm.hide).to be_nil
-    end
-  end
-
   describe "#serializer" do
     it "returns provided :serializer option" do
       expect(normalizer.new(opts: {serializer: 123}).serializer).to eq 123
@@ -137,6 +119,149 @@ RSpec.describe Serega::SeregaAttributeNormalizer do
       expect(block).to be_a Proc
       expect(block.call(object1)).to be_nil
       expect(block.call(object2)).to eq "NAME"
+    end
+  end
+
+  describe "preloads functionality" do
+    let(:initials) { {name: :foo, opts: opts, block: nil} }
+    let(:opts) { {} }
+    let(:norm) { normalizer.new(**initials) }
+
+    describe "#preloads" do
+      it "returns empty hash for regular attributes" do
+        expect(norm.preloads).to eq({})
+      end
+
+      it "returns nil when provided nil" do
+        opts[:preload] = nil
+        expect(norm.preloads).to be_nil
+      end
+
+      it "returns formatted provided preloads" do
+        opts[:preload] = :bar
+        expect(norm.preloads).to eq(bar: {})
+        expect(norm.preloads).to equal norm.preloads
+      end
+
+      it "returns automatically found preloads when serializer provided" do
+        serializer_class.config.auto_preload_attributes_with_serializer = true
+        opts[:serializer] = "bar"
+        expect(norm.preloads).to eq(foo: {})
+      end
+
+      it "returns no preloads for attributes with serializer by default" do
+        opts[:serializer] = "bar"
+        expect(norm.preloads).to eq({})
+      end
+
+      it "returns automatically found preloads when :delegate option provided" do
+        serializer_class.config.auto_preload_attributes_with_delegate = true
+        opts[:delegate] = {to: :bar}
+        expect(norm.preloads).to eq(bar: {})
+      end
+
+      it "returns no preloads for attributes with :delegate option by default" do
+        opts[:delegate] = {to: :bar}
+        expect(norm.preloads).to eq({})
+      end
+    end
+
+    describe "#preloads_path" do
+      it "returns constructed preloads_path" do
+        opts[:preload] = :foo
+        expect(norm.preloads_path).to eq([:foo])
+        expect(norm.preloads_path).to be_frozen
+        expect(norm.preloads_path).to equal norm.preloads_path
+      end
+
+      it "returns provided preloads_path" do
+        opts[:serializer] = "foo"
+        opts[:preload] = %i[bar bazz]
+        opts[:preload_path] = :bar
+        expect(norm.preloads_path).to eq([:bar])
+        expect(norm.preloads_path).to be_frozen
+      end
+
+      it "returns symbolized preloads_path" do
+        opts[:serializer] = "foo"
+        opts[:preload] = %i[bar bazz]
+        opts[:preload_path] = "bar"
+        expect(norm.preloads_path).to eq([:bar])
+      end
+
+      it "returns symbolized preloads_path when array provided" do
+        opts[:serializer] = "foo"
+        opts[:preload] = %i[bar bazz]
+        opts[:preload_path] = %w[bar]
+        expect(norm.preloads_path).to eq([:bar])
+      end
+
+      it "returns normalized array of provided paths" do
+        opts[:serializer] = "foo"
+        opts[:preload] = %i[bar bazz]
+        opts[:preload_path] = [["bar"], ["bar", :bazz]]
+        expect(norm.preloads_path).to eq([%i[bar], %i[bar bazz]])
+        expect(norm.preloads_path).to be_frozen
+      end
+
+      it "returns nil if nil provided" do
+        opts[:serializer] = "foo"
+        opts[:preload] = %i[bar bazz]
+        opts[:preload_path] = nil
+        expect(norm.preloads_path).to be_nil
+      end
+    end
+
+    describe "#hide with preloads" do
+      context "without auto_hide config" do
+        it "returns opt :hide" do
+          a0 = serializer_class.attribute :a0
+          a1 = serializer_class.attribute :a1, preload: :a1
+          a2 = serializer_class.attribute :a2, preload: :a2, hide: true
+          a3 = serializer_class.attribute :a3, preload: :a3, hide: false
+
+          expect(a0.hide).to be_nil
+          expect(a1.hide).to be_nil
+          expect(a2.hide).to be true
+          expect(a3.hide).to be false
+        end
+
+        it "returns saved :hide option" do
+          norm = normalizer.new(opts: {hide: true})
+
+          expect(norm.hide).to be true
+          expect(norm.instance_variable_get(:@hide)).to be true
+
+          norm.instance_variable_set(:@hide, nil)
+          expect(norm.hide).to be_nil
+        end
+      end
+
+      context "with auto_hide config" do
+        before do
+          serializer_class.config.auto_hide_attributes_with_preload = true
+        end
+
+        it "returns opt :hide => true when preload is not blank" do
+          a0 = serializer_class.attribute :a0
+          a1 = serializer_class.attribute :a1, preload: :a1
+          a2 = serializer_class.attribute :a2, preload: :a2, hide: true
+          a3 = serializer_class.attribute :a3, preload: :a3, hide: false
+          a4 = serializer_class.attribute :a4, preload: nil
+          a5 = serializer_class.attribute :a5, preload: false
+          a6 = serializer_class.attribute :a6, preload: {}
+          a7 = serializer_class.attribute :a7, preload: []
+
+          expect(a0.hide).to be_nil
+          expect(a1.hide).to be true
+          expect(a2.hide).to be true
+          expect(a3.hide).to be false
+          expect(a4.hide).to be_nil
+          expect(a5.hide).to be_nil
+          expect(a6.hide).to be_nil
+          expect(a7.hide).to be_nil
+        end
+      end
     end
   end
 end