+++ spec/serega_spec.rb
@@ -20,6 +20,9 @@ RSpec.describe Serega do
         check_initiate_params
         delegate_default_allow_nil
         max_cached_plans_per_serializer_count
+        auto_preload_attributes_with_delegate
+        auto_preload_attributes_with_serializer
+        auto_hide_attributes_with_preload
         to_json
         from_json
       ]
@@ -27,7 +30,7 @@ RSpec.describe Serega do
       expect(config.plugins).to eq []
       expect(config.serialize_keys).to match_array(%i[context many])
       expect(config.initiate_keys).to match_array(%i[only except with check_initiate_params])
-      expect(config.attribute_keys).to match_array(%i[method value serializer many hide const delegate default])
+      expect(config.attribute_keys).to match_array(%i[method value serializer many hide const delegate default preload preload_path])
       expect(config.check_attribute_name).to be true
       expect(config.check_initiate_params).to be true
       expect(config.delegate_default_allow_nil).to be false
@@ -612,6 +615,123 @@ RSpec.describe Serega do
 
       it "returns hash with only requested fields and all fields of requested relation" do
         expect(result).to eq({comment: {text: "TEXT"}})
+      end
+    end
+  end
+
+  describe "Preloads functionality" do
+    let(:serializer_class) { Class.new(described_class) }
+
+    describe "configuration" do
+      it "has preload options in attribute_keys by default" do
+        attribute_keys = serializer_class.config.attribute_keys
+        expect(attribute_keys).to include :preload
+        expect(attribute_keys).to include :preload_path
+      end
+
+      it "configures to not preload attributes with serializer by default" do
+        expect(serializer_class.config.auto_preload_attributes_with_serializer).to be false
+      end
+
+      it "configures to not preload attributes with :delegate option by default" do
+        expect(serializer_class.config.auto_preload_attributes_with_delegate).to be false
+      end
+
+      it "configures to not hide attributes with preload option by default" do
+        expect(serializer_class.config.auto_hide_attributes_with_preload).to be false
+      end
+
+      it "allows to configure to preload attributes with serializer" do
+        serializer_class.config.auto_preload_attributes_with_serializer = true
+        expect(serializer_class.config.auto_preload_attributes_with_serializer).to be true
+      end
+
+      it "allows to configure to preload attributes with :delegate option" do
+        serializer_class.config.auto_preload_attributes_with_delegate = true
+        expect(serializer_class.config.auto_preload_attributes_with_delegate).to be true
+      end
+
+      it "allows to configure to hide attributes with preloads" do
+        serializer_class.config.auto_hide_attributes_with_preload = true
+        expect(serializer_class.config.auto_hide_attributes_with_preload).to be true
+      end
+
+      it "raises error when not boolean value provided to auto_preload_attributes_with_serializer" do
+        expect { serializer_class.config.auto_preload_attributes_with_serializer = nil }
+          .to raise_error Serega::SeregaError, "Must have boolean value, #{nil.inspect} provided"
+      end
+
+      it "raises error when not boolean value provided to auto_preload_attributes_with_delegate" do
+        expect { serializer_class.config.auto_preload_attributes_with_delegate = nil }
+          .to raise_error Serega::SeregaError, "Must have boolean value, #{nil.inspect} provided"
+      end
+
+      it "raises error when not boolean value provided to auto_hide_attributes_with_preload" do
+        expect { serializer_class.config.auto_hide_attributes_with_preload = nil }
+          .to raise_error Serega::SeregaError, "Must have boolean value, #{nil.inspect} provided"
+      end
+    end
+
+    describe "instance methods" do
+      it "adds #preloads method" do
+        serializer = serializer_class.new
+        expect(serializer).to respond_to(:preloads)
+      end
+
+      it "returns empty hash when no preloads defined" do
+        serializer_class.attribute :name
+        serializer = serializer_class.new
+        expect(serializer.preloads).to eq({})
+      end
+
+      it "returns preloads hash when preloads defined" do
+        serializer_class.attribute :name, preload: :user_profile
+        serializer = serializer_class.new
+        expect(serializer.preloads).to eq({user_profile: {}})
+      end
+    end
+
+    describe "attribute preloads" do
+      it "allows manual preload specification" do
+        serializer_class.attribute :name, preload: :user_profile
+        attribute = serializer_class.attributes[:name]
+        expect(attribute.preloads).to eq({user_profile: {}})
+      end
+
+      it "auto-preloads for delegate when configured" do
+        serializer_class.config.auto_preload_attributes_with_delegate = true
+        serializer_class.attribute :name, delegate: {to: :profile}
+        attribute = serializer_class.attributes[:name]
+        expect(attribute.preloads).to eq({profile: {}})
+      end
+
+      it "auto-preloads for serializer when configured" do
+        other_serializer = Class.new(described_class)
+        serializer_class.config.auto_preload_attributes_with_serializer = true
+        serializer_class.attribute :profile, serializer: other_serializer
+        attribute = serializer_class.attributes[:profile]
+        expect(attribute.preloads).to eq({profile: {}})
+      end
+
+      it "auto-hides attributes with preloads when configured" do
+        serializer_class.config.auto_hide_attributes_with_preload = true
+        serializer_class.attribute :name, preload: :user_profile
+        attribute = serializer_class.attributes[:name]
+        expect(attribute.hide).to be true
+      end
+
+      it "allows disabling preloads with false" do
+        serializer_class.attribute :name, preload: false
+        attribute = serializer_class.attributes[:name]
+        expect(attribute.preloads).to be_nil
+      end
+    end
+
+    describe "validation" do
+      it "validates preload option cannot be used with const" do
+        expect {
+          serializer_class.attribute :name, preload: :profile, const: "value"
+        }.to raise_error Serega::SeregaError, "Option :preload can not be used together with option :const"
       end
     end
   end